chore：升级jodconverter版本至4.4.0，jodconverter-local升级-》4.4.8

+++ y9-module-jodconverter/risenet-y9boot-webapp-jodconverter/src/main/java/net/risesoft/service/impl/CadFilePreviewImpl.java
@@ -15,6 +15,7 @@ import net.risesoft.service.FileHandlerService;
 import net.risesoft.service.FilePreview;
 import net.risesoft.utils.DownloadUtils;
 import net.risesoft.utils.KkFileUtils;
+import net.risesoft.utils.WebUtils;
 import net.risesoft.web.filter.BaseUrlFilter;
 
 @Slf4j
@@ -23,7 +24,6 @@ public class CadFilePreviewImpl implements FilePreview {
 
     private static final String OFFICE_PREVIEW_TYPE_IMAGE = "image";
     private static final String OFFICE_PREVIEW_TYPE_ALL_IMAGES = "allImages";
-    private static final String FILE_DIR = ConfigConstants.getFileDir();
 
     private final FileHandlerService fileHandlerService;
     private final OtherFilePreviewImpl otherFilePreview;
@@ -41,53 +41,51 @@ public class CadFilePreviewImpl implements FilePreview {
         String baseUrl = BaseUrlFilter.getBaseUrl();
         boolean forceUpdatedCache = fileAttribute.forceUpdatedCache();
         String fileName = fileAttribute.getName();
-        String suffix = fileAttribute.getSuffix();
         String cadPreviewType = ConfigConstants.getCadPreviewType();
-        String pdfName = fileName.substring(0, fileName.lastIndexOf(".")) + suffix + "." + cadPreviewType; // 生成文件添加类型后缀
-                                                                                                           // 防止同名文件
-        String outFilePath = FILE_DIR + pdfName;
+        String cacheName = fileAttribute.getCacheName();
+        String outFilePath = fileAttribute.getOutFilePath();
         // 判断之前是否已转换过，如果转换过，直接返回，否则执行转换
-        if (forceUpdatedCache || !fileHandlerService.listConvertedFiles().containsKey(pdfName)
+        if (forceUpdatedCache || !fileHandlerService.listConvertedFiles().containsKey(cacheName)
             || !ConfigConstants.isCacheEnabled()) {
-            String filePath;
-            ReturnResponse<String> response = DownloadUtils.downLoad(fileAttribute, null);
+            ReturnResponse<String> response = DownloadUtils.downLoad(fileAttribute, fileName);
             if (response.isFailure()) {
                 return otherFilePreview.notSupportedFile(model, fileAttribute, response.getMsg());
             }
-            filePath = response.getContent();
+            String filePath = response.getContent();
             String imageUrls = null;
             if (StringUtils.hasText(outFilePath)) {
                 try {
-                    imageUrls = fileHandlerService.cadToPdf(filePath, outFilePath, cadPreviewType);
+                    imageUrls = fileHandlerService.cadToPdf(filePath, outFilePath, cadPreviewType, fileAttribute);
                 } catch (Exception e) {
                     LOGGER.error("cad 转 pdf 异常！", e);
                 }
                 if (imageUrls == null) {
-                    return otherFilePreview.notSupportedFile(model, fileAttribute, "office转图片异常，请联系管理员");
+                    return otherFilePreview.notSupportedFile(model, fileAttribute, "CAD转换异常，请联系管理员");
                 }
                 // 是否保留CAD源文件
-                if (ConfigConstants.getDeleteSourceFile()) {
+                if (!fileAttribute.isCompressFile() && ConfigConstants.getDeleteSourceFile()) {
                     KkFileUtils.deleteFileByPath(filePath);
                 }
                 if (ConfigConstants.isCacheEnabled()) {
                     // 加入缓存
-                    fileHandlerService.addConvertedFile(pdfName, fileHandlerService.getRelativePath(outFilePath));
+                    fileHandlerService.addConvertedFile(cacheName, fileHandlerService.getRelativePath(outFilePath));
                 }
             }
         }
+        cacheName = WebUtils.encodeFileName(cacheName);
         if ("tif".equalsIgnoreCase(cadPreviewType)) {
-            model.addAttribute("currentUrl", pdfName);
+            model.addAttribute("currentUrl", cacheName);
             return TIFF_FILE_PREVIEW_PAGE;
         } else if ("svg".equalsIgnoreCase(cadPreviewType)) {
-            model.addAttribute("currentUrl", pdfName);
+            model.addAttribute("currentUrl", cacheName);
             return SVG_FILE_PREVIEW_PAGE;
         }
         if (baseUrl != null && (OFFICE_PREVIEW_TYPE_IMAGE.equals(officePreviewType)
             || OFFICE_PREVIEW_TYPE_ALL_IMAGES.equals(officePreviewType))) {
-            return getPreviewType(model, fileAttribute, officePreviewType, baseUrl, pdfName, outFilePath,
-                fileHandlerService, OFFICE_PREVIEW_TYPE_IMAGE, otherFilePreview);
+            return getPreviewType(model, fileAttribute, officePreviewType, cacheName, outFilePath, fileHandlerService,
+                OFFICE_PREVIEW_TYPE_IMAGE, otherFilePreview);
         }
-        model.addAttribute("pdfUrl", pdfName);
+        model.addAttribute("pdfUrl", cacheName);
         return PDF_FILE_PREVIEW_PAGE;
     }
 }
